🐛Fix: dynamic modal error handleing
