+++ src/views/create-exhibition/ui/createExhibition/index.tsx
@@ -6,7 +6,7 @@ const CreateExhibition = () => {
   return (
     <div className="flex h-screen flex-col gap-[30px] mobile:gap-0">
       <Header />
-      <div className="mx-auto w-full max-w-[792px] flex-1 px-5">
+      <div className="mx-auto w-full max-w-[792px] flex-1 px-5 pb-5">
         <ExhibitionForm />
       </div>
     </div>
